Add files via upload
The last 4 versions of the game. Allows multi-level play: Beginner (3 numbers), Medium (4 numbers), Advanced (5) and Expert (6 numbers). Retrieves registered users and creates new users in a simple JSON database. Allows to play as a "guest" if the player does not want to register. The database is portable with the game, so statistics can be continued on any device.